PR #3 changes
