+++ CraftingRevisions/Patches/GameManager_Awake.cs
@@ -1,5 +1,7 @@
 using HarmonyLib;
 using Il2Cpp;
+using System.Reflection;
+using UnityEngine;
 
 namespace CraftingRevisions.Patches
 {
@@ -22,7 +24,6 @@ namespace CraftingRevisions.Patches
 	{
 		private static void Postfix(Il2CppTLD.Gear.BlueprintManager __instance)
 		{
-
 			// check we have a HasSet and it has contents
 			if (BlueprintManager.jsonUserBlueprints != null && BlueprintManager.jsonUserBlueprints.Count > 0)
 			{
@@ -35,8 +36,8 @@ namespace CraftingRevisions.Patches
 
 					if (loaded == false)
 					{
-						// WIP
-						//BlueprintManager.ValidateJsonBlueprint(jsonUserBlueprint);
+						// validate the blueprint
+						BlueprintManager.ValidateJsonBlueprint(jsonUserBlueprint);
 					}
 				}
 			}
@@ -49,15 +50,50 @@ namespace CraftingRevisions.Patches
 	[HarmonyPatch(typeof(Il2CppTLD.Gear.UserBlueprintData), nameof(Il2CppTLD.Gear.UserBlueprintData.MakeRuntimeWwiseEvent), new Type[] { typeof(string) })]
 	internal class UserBlueprintData_MakeRuntimeWwiseEvent
 	{
-		private static void Prefix(string eventName, ref bool __runOriginal)
+		private static void Prefix(ref bool __runOriginal)
 		{
 			__runOriginal = false;
 		}
 		private static void Postfix(string eventName, ref bool __runOriginal, ref Il2CppAK.Wwise.Event? __result)
 		{
-			__result = BlueprintManager.MakeAudioEvent(eventName);
+			__result = MakeAudioEvent(eventName);
 			__runOriginal = false;
 		}
+
+		internal static Il2CppAK.Wwise.Event? MakeAudioEvent(string? eventName)
+		{
+			Il2CppAK.Wwise.Event emptyEvent = new();
+			emptyEvent.WwiseObjectReference = ScriptableObject.CreateInstance<WwiseEventReference>();
+
+			if (eventName == null || eventName == "")
+			{
+				return emptyEvent;
+			}
+			uint eventId = GetAKEventIdFromString(eventName);
+			if (eventId == 0U)
+			{
+				return emptyEvent;
+			}
+
+			Il2CppAK.Wwise.Event newEvent = new();
+			newEvent.WwiseObjectReference = ScriptableObject.CreateInstance<WwiseEventReference>();
+			newEvent.WwiseObjectReference.objectName = eventName;
+			newEvent.WwiseObjectReference.id = eventId;
+			return newEvent;
+		}
+		internal static uint GetAKEventIdFromString(string eventName)
+		{
+			Type type = typeof(Il2CppAK.EVENTS);
+			foreach (PropertyInfo prop in type.GetProperties(BindingFlags.Static | BindingFlags.Public | BindingFlags.Instance))
+			{
+				if (prop.Name.ToLower() == eventName.ToLower())
+				{
+					return Convert.ToUInt32(prop.GetValue(null)?.ToString());
+				}
+			}
+			return 0U;
+		}
+
 	}
 
 }
